base+flowable

+++ campus-assistance-common/src/main/java/com/siyu/common/domain/PaginationQuery.java
@@ -3,17 +3,16 @@ package com.siyu.common.domain;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
 public class PaginationQuery<T> {
 
     @ApiModelProperty(value = "页码", example = "1")
-    private Long pageNum;
+    private Integer pageNum;
 
     @ApiModelProperty(value = "每页记录", example = "10")
-    private Long pageSize;
+    private Integer pageSize;
 
     @ApiModelProperty(value = "查询条件")
     private T condition;
